New category with one application added
It was to hard to decide if load testing should be in testing or profiling, hence new category.

Tsung itself has two pages. [First](http://www.process-one.net/en/tsung/), to which github repo points, is not updated to often. [Second](http://tsung.erlang-projects.org/index.en.html), which seems to be up to date, does not look too impressive.  So I decided to go with link to github repo.